+++ src/framework/piral-core/src/components/ExtensionSlot.tsx
@@ -9,12 +9,8 @@ import { ExtensionSlotProps } from '../types';
  * extensions of a given name should be rendered at a specific
  * location.
  */
-export function ExtensionSlot<T extends string>({
-  name,
-  render = defaultRender,
-  empty,
-  params,
-}: ExtensionSlotProps<T>) {
+export function ExtensionSlot<T extends string>(props: ExtensionSlotProps<T>) {
+  const { name, render = defaultRender, empty, params } = props;
   const extensions = useGlobalState(s => s.registry.extensions[name] || []);
   return render(
     extensions.length === 0 && isfunc(empty)