+++ src/tooling/piral-cli/src/common/constants.ts
@@ -1,4 +1,3 @@
-export const defaultCacheDir = '.cache';
 export const defaultRegistry = 'https://registry.npmjs.org/';
 export const filesTar = 'files';
 export const filesOnceTar = 'files_once';